requirements.txt
streamlit==1.28.0
spacy==3.7.2
python-docx==0.8.11
PyPDF2==3.0.1
pdfminer.six==20240720
language-tool-python==2.8
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.2
